Modify schemas for result (method, params) (#76)
They are now `expected` to be a list with a single str value
(method) and a dict with a single entry with the same name
as the method (params). Note that so far this is not enforced,
checks need to be set in place to make sure there is only
one item in each of these.