Adapt schemas for serialization limitations
Also changed the types from the results so that methods would no longer
be just a string, but a list of string of expected length 1 (we should
re-think this in the future). Also the parameters are not just a dict,
but a nested dict with a single outermost key '<method_name>' pointing
to a value that has the actual parameters.