Add basic limitations mechanism (#127)
The idea is the limitations will basically be sub-schemas from the capabilities.
So if every (quantity, method) pair has a `capability` schema that defines
what each field of its inputs (and outputs) of the method means, then each
(tenant, quantity, method) triad has a sub-schema of that which defines
things such as ranges for the numbers or possible values for the strings.

So far the functionalities available are:

- A command to add tenants to the tenant table (specifying the limitations).
The limitations provided are validated as a compatible sub-schema of the
corresponding (quantity, method) by using the jsonsubschema package.
As a note, this package seems a bit limited, I need to check it a bit further
but it is unclear to me at this point if there is a way to get any info of *why*
a given schema may not be a sub-schema of the reference schema.

- An endpoint get all limitations, provided as a list of (quantity, method,
limitations). Each combination of (quantity, method) should appear only
once in the list: the limitations for all tenant that offer the same quantity+
method are combined using the "anyOf" schema descriptor.

So far this limitations are just informative to the users and are not enforced
on the requests (or results), but they can use them to validate their requests
before submitting. A next step could be to have the server check that
submitted requests are valid for the limitations, with an optional flag during
submission to ignore it (by default we reject a temperature of 1000K if no
current tenant supports it, but the submitter can choose to force the server
to accept the request anyways).