FIX: Externalize configuration file from codebase (#131)
The configuration was being directly imported from a python
file that users were expected to create and set up inside of
the codebase. Now it is taken from a json file whose location
can be configured using the `FINALES_CONFIG_FILEPATH`
environment variable, but that defaults to a `config.json`
file inside the `src/FINALES2/config` folder.