Redo limitations mechanism
Returning to the simplified limitations that are a modification of the
parameters (with lists for options instead of just values), as opposed
to them being sub-schemas of the capability schema. Although the
sub-schema implementation is more versatile with less effort, manually
building schemas is quite complex, so we reverted it back.