BREAK #157: Change schemas returned by API
The API was returning `Result / Request` and `ResultInfo / RequestInfo`
objects in an inconsistent way. The difference between these is that the
first contain just the "scientific" info exchanged, while the `Info` objects
also contain metadata related to the internal handling of FINALES (like
the uuid of the message, timestamp, status, etc). Before we were not
sure about exposing this information unnecessarily, but now that we
want users to have access to the status of the requests and results
we decided to just expose all the information for consistency.

THIS COMMIT CONTAINS BREAKING CHANGES IN THE API