FIX problem with unnecessary `allOf` keyword.
For some reason, some of the schemas generated contain the
"allOf" keyword with a single list element instead of just
describing that element.

Since we didn't contemplate the use of `allOf` for how we
were designing the schemas, this was breaking the registration
of tenants because the json schemas for the limitation could
not be built.

Now a clause was added to detect this specific case and take
out the unnecessary `allOf`.