ADD to make limitations easier to create
Now the limitations for a schema can be either an instance of an
object (if that is the only possibility) or an array given by the
type of the schema.

This prevents unnecessary wrapping around lists, but may complicate
later the unification of limitations or the implementation of
validation checks.