Bugfix verification of submissions
When updating a schema for a capability, FINALES validated
submissions against the first schema for the respective
quantity and method found in the database, which is the
first schema that was registered for the respective capability.
This behaviour leads to issues when a schema is updated, because an
error occurrs due to the validation failing for the old inactive
schema. The improvement now checks first against the currently
active schema and subsequently agains the latest inactive one.